Massive update

+++ src/TeeKay87.MemoryEngine.Core/Scanning/MemoryScanProgress.cs
@@ -3,4 +3,4 @@ namespace TeeKay87.MemoryEngine.Core.Scanning;
 public sealed record MemoryScanProgress(
     ulong UnitsProcessed,
     ulong UnitsTotal,
-    int ResultsFound);
+    long ResultsFound);
